# Lammah-game-backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ai-agent/generate-questions": {
      "post": {
        "operationId": "aiGenerateQuestions",
        "summary": "Generate questions for a category using the configured AI provider",
        "description": "Generates open-answer questions for a selected category. Questions are saved as drafts and must be reviewed/approved by admin before use in games.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQuestionsDto"
              },
              "examples": {
                "default": {
                  "summary": "Generate questions for category",
                  "value": {
                    "categoryId": "66b8f5f2c9d7a8b1e3f00222",
                    "count": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Questions generated and saved as drafts successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateQuestionsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - category not found or validation error"
          },
          "500": {
            "description": "Internal server error - AI provider API error or other issues"
          }
        },
        "tags": [
          "AI Agent"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/ai-generator/debug-tools": {
      "get": {
        "operationId": "aiDebugTools",
        "summary": "Debug local media tool availability for AI asset generation",
        "deprecated": true,
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiToolDiagnosticsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin AI Generator"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/ai-generator/generate-reviewed": {
      "post": {
        "operationId": "aiGenerateReviewed",
        "summary": "Generate reviewed AI question drafts without saving them",
        "description": "Generates multiple-choice question drafts from the configured AI provider, validates quality issues, and returns drafts only.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReviewedQuestionsDto"
              },
              "examples": {
                "default": {
                  "summary": "Generate reviewed Arabic World Cup questions",
                  "value": {
                    "catalogName": "رياضة",
                    "categoryName": "كأس العالم",
                    "difficulty": "medium",
                    "count": 10,
                    "language": "ar"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reviewed question drafts generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateReviewedQuestionsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin AI Generator"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/ai-generator/save-drafts": {
      "post": {
        "operationId": "aiSaveReviewedDrafts",
        "summary": "Persist reviewed questions as unpublished AI drafts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveReviewedDraftsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveReviewedDraftsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin AI Generator"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Lammah Game API",
    "description": "Runtime HTTP contract for Lammah game administration, content, and gameplay.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DifficultyLevel": {
        "type": "string",
        "enum": [
          "easy",
          "medium",
          "hard"
        ]
      },
      "GameMode": {
        "type": "string",
        "enum": [
          "trivia",
          "identifyCharacter",
          "identifyVoice",
          "identifyImage",
          "completeQuote",
          "timeline",
          "emojiPuzzle",
          "identifySong",
          "identifySinger",
          "identifyMusicIntro"
        ]
      },
      "QuestionStatus": {
        "type": "string",
        "enum": [
          "draft",
          "approved",
          "published",
          "archived",
          "rejected"
        ]
      },
      "QuestionSource": {
        "type": "string",
        "enum": [
          "manual",
          "ai",
          "imported",
          "music"
        ]
      },
      "QuestionAssetType": {
        "type": "string",
        "enum": [
          "audio",
          "image",
          "video",
          "gif"
        ]
      },
      "QuestionAssetResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/QuestionAssetType"
          },
          "url": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string"
          },
          "searchQuery": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "type",
          "url",
          "source"
        ]
      },
      "AssetStatus": {
        "type": "string",
        "enum": [
          "NOT_REQUIRED",
          "PENDING",
          "READY",
          "FAILED"
        ]
      },
      "QuestionResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "answer": {
            "type": "string"
          },
          "correctAnswer": {
            "type": "string"
          },
          "wrongAnswers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "explanation": {
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/DifficultyLevel"
          },
          "points": {
            "type": "number",
            "enum": [
              200,
              400,
              600
            ]
          },
          "score": {
            "type": "number",
            "enum": [
              200,
              400,
              600
            ]
          },
          "gameMode": {
            "$ref": "#/components/schemas/GameMode"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionStatus"
          },
          "source": {
            "$ref": "#/components/schemas/QuestionSource"
          },
          "primaryAsset": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionAssetResponseDto"
              }
            ]
          },
          "coverImage": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionAssetResponseDto"
              }
            ]
          },
          "primaryAssetRequest": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "coverImageRequest": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "coverImageStatus": {
            "$ref": "#/components/schemas/AssetStatus"
          },
          "coverImageFailureReason": {
            "type": "string"
          },
          "assetStatus": {
            "$ref": "#/components/schemas/AssetStatus"
          },
          "assetFailureReason": {
            "type": "string"
          },
          "assetFailureStep": {
            "type": "string"
          },
          "assetFailureDiagnostics": {
            "type": "object",
            "additionalProperties": true
          },
          "gameplayMetadata": {
            "type": "object",
            "additionalProperties": true
          },
          "aiMetadata": {
            "type": "object",
            "additionalProperties": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "mediaUrl": {
            "type": "string"
          },
          "hasPreviewAudio": {
            "type": "boolean"
          },
          "isFreeGameQuestion": {
            "type": "boolean"
          },
          "qualityScore": {
            "type": "number"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "_id",
          "question",
          "wrongAnswers",
          "difficulty",
          "status",
          "source"
        ]
      },
...
      "GenerateQuestionsDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "count": {
            "type": "number",
            "minimum": 1,
            "maximum": 20,
            "default": 2
          }
        },
        "required": [
          "categoryId"
        ]
      },
      "GenerateQuestionsResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "count": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionResponseDto"
            }
          }
        },
        "required": [
          "statusCode",
          "message",
          "count",
          "data"
        ]
      },
      "AiToolDiagnosticsResponseDto": {
        "type": "object",
        "properties": {
          "ffmpegAvailable": {
            "type": "boolean"
          },
          "ytDlpAvailable": {
            "type": "boolean"
          },
          "ffmpegVersion": {
            "type": "string"
          },
          "ytDlpVersion": {
            "type": "string"
          }
        },
        "required": [
          "ffmpegAvailable",
          "ytDlpAvailable",
          "ffmpegVersion",
          "ytDlpVersion"
        ]
      },
      "GenerateReviewedQuestionsDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "catalogName": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "count": {
            "type": "number",
            "minimum": 1,
            "maximum": 20,
            "default": 2
          },
          "language": {
            "type": "string",
            "enum": [
              "ar"
            ]
          }
        }
      },
      "AiAssetRequestResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "quote",
              "emoji",
              "timeline",
              "video",
              "gif"
            ]
          },
          "assetType": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "quote",
              "emoji",
              "timeline",
              "video",
              "gif"
            ]
          },
          "provider": {
            "type": "string"
          },
          "query": {
            "type": "string"
          },
          "entity": {
            "type": "string"
          },
          "franchise": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "originalName": {
            "type": "string"
          },
          "localizedName": {
            "type": "string"
          },
          "englishTitle": {
            "type": "string"
          },
          "arabicTitle": {
            "type": "string"
          },
          "context": {
            "type": "string"
          },
          "entityType": {
            "type": "string"
          },
          "visualHint": {
            "type": "string"
          },
          "categoryType": {
            "type": "string"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "gameplay",
              "decorative"
            ]
          },
          "duration": {
            "type": "number"
          },
          "speaker": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ]
      },
      "AiAssetResponseDto": {
        "type": "object",
        "properties": {
          "localPath": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "quote",
              "emoji",
              "timeline",
              "video",
              "gif"
            ]
          },
          "duration": {
            "type": "number"
          },
          "sourceUrl": {
            "type": "string"
          },
          "searchQuery": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "localPath",
          "url",
          "source",
          "provider",
          "type"
        ]
      },
      "AiAgentTraceResponseDto": {
        "type": "object",
        "properties": {
          "agent": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "failed",
              "fallback"
            ]
          },
          "durationMs": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "agent",
          "status",
          "durationMs"
        ]
      },
      "ReviewedQuestionDraftResponseDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          },
          "correctAnswer": {
            "type": "string"
          },
          "wrongAnswers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "gameMode": {
            "type": "string",
            "enum": [
              "trivia",
              "identifyCharacter",
              "identifyVoice",
              "identifyImage",
              "completeQuote",
              "timeline",
              "emojiPuzzle",
              "identifySong",
              "identifySinger",
              "identifyMusicIntro"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "quote",
              "emoji",
              "timeline",
              "video",
              "gif"
            ]
          },
          "assetRequest": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AiAssetRequestResponseDto"
              }
            ]
          },
          "assetStatus": {
            "type": "string",
            "enum": [
              "NOT_REQUIRED",
              "PENDING",
              "READY",
              "FAILED"
            ]
          },
          "asset": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AiAssetResponseDto"
              }
            ]
          },
          "primaryAssetRequest": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AiAssetRequestResponseDto"
              }
            ]
          },
          "primaryAssetStatus": {
            "type": "string",
            "enum": [
              "NOT_REQUIRED",
              "PENDING",
              "READY",
              "FAILED"
            ]
          },
          "primaryAsset": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AiAssetResponseDto"
              }
            ]
          },
          "coverImageRequest": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AiAssetRequestResponseDto"
              }
            ]
          },
          "coverImageStatus": {
            "type": "string",
            "enum": [
              "NOT_REQUIRED",
              "PENDING",
              "READY",
              "FAILED"
            ]
          },
          "coverImage": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AiAssetResponseDto"
              }
            ]
          },
          "coverImageFailureReason": {
            "type": "string",
            "nullable": true
          },
          "assetFailureReason": {
            "type": "string"
          },
          "assetFailureStep": {
            "type": "string"
          },
          "assetFailureDiagnostics": {
            "oneOf": [
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            ]
          },
          "wasGameplayAutoFixed": {
            "type": "boolean"
          },
          "gameplayFixReason": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "qualityScore": {
            "type": "number"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "agentTrace": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AiAgentTraceResponseDto"
            }
          },
          "gameplayMetadata": {
            "type": "object",
            "additionalProperties": true
          },
          "aiMetadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "question",
          "correctAnswer",
          "wrongAnswers",
          "difficulty",
          "gameMode",
          "type",
          "assetStatus",
          "primaryAssetStatus",
          "coverImageStatus",
          "explanation",
          "qualityScore",
          "issues"
        ]
      },
      "ReviewedQuestionsDataResponseDto": {
        "type": "object",
        "properties": {
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewedQuestionDraftResponseDto"
            }
          }
        },
        "required": [
          "questions"
        ]
      },
      "GenerateReviewedQuestionsResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 201
          },
          "message": {
            "type": "string"
          },
          "count": {
            "type": "number"
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          },
          "data": {
            "$ref": "#/components/schemas/ReviewedQuestionsDataResponseDto"
          }
        },
        "required": [
          "statusCode",
          "message",
          "count",
          "meta",
          "data"
        ]
      },
      "SaveReviewedDraftsDto": {
        "type": "object",
        "properties": {
          "drafts": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "categoryId": {
            "type": "string"
          },
          "catalogId": {
            "type": "string"
          }
        },
        "required": [
          "drafts",
          "categoryId"
        ]
      },
      "SaveDraftFailureDto": {
        "type": "object",
        "properties": {
          "index": {
            "type": "number",
            "minimum": 0
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "index",
          "reason"
        ]
      },
      "SaveReviewedDraftsResponseDto": {
        "type": "object",
        "properties": {
          "savedCount": {
            "type": "number",
            "minimum": 0
          },
          "failedCount": {
            "type": "number",
            "minimum": 0
          },
          "savedQuestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionResponseDto"
            }
          },
          "failures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaveDraftFailureDto"
            }
          }
        },
        "required": [
          "savedCount",
          "failedCount",
          "savedQuestions",
          "failures"
        ]
      },
...
    }
  }
}