# Lammah-game-backend/src/modules/ai-agent/admin-ai-generator.controller.ts
import {
  Body,
  Controller,
  Get,
  HttpStatus,
  Post,
  UseGuards,
} from '@nestjs/common';
import {
  ApiBearerAuth,
  ApiBody,
  ApiOperation,
  ApiResponse,
  ApiTags,
} from '@nestjs/swagger';
import { execFile } from 'child_process';
import { promisify } from 'util';
import { Roles } from '../../common/decorators/roles.decorator';
import { JwtAuthGuard } from '../../common/guards/jwt-auth.guard';
import { RolesGuard } from '../../common/guards/roles.guard';
import { UserRole } from '../users/schemas/user.schema';
import { AiAgentService } from './ai-agent.service';
import { GenerateReviewedQuestionsDto } from './dto/generate-reviewed-questions.dto';
import { SaveReviewedDraftsDto } from './dto/save-reviewed-drafts.dto';
import {
  AiToolDiagnosticsResponseDto,
  GenerateReviewedQuestionsResponseDto,
  SaveReviewedDraftsResponseDto,
} from './dto/ai-response.dto';
import { QuestionResponseMapper } from '../questions/mappers/question-response.mapper';

const execFileAsync = promisify(execFile);

@ApiTags('Admin AI Generator')
@ApiBearerAuth()
@Controller('admin/ai-generator')
@UseGuards(JwtAuthGuard, RolesGuard)
@Roles(UserRole.ADMIN)
export class AdminAiGeneratorController {
  constructor(private readonly aiAgentService: AiAgentService) {}

  @Get('debug-tools')
  @ApiOperation({
    operationId: 'aiDebugTools',
    summary: 'Debug local media tool availability for AI asset generation',
    deprecated: true,
  })
  @ApiResponse({ status: 200, type: AiToolDiagnosticsResponseDto })
  async debugTools() {
    const [ffmpegWhich, ytDlpWhich, ffmpegVersion, ytDlpVersion] =
      await Promise.all([
        this.runToolCommand('which', ['ffmpeg']),
        this.runToolCommand('which', ['yt-dlp']),
        this.runToolCommand('ffmpeg', ['-version']),
        this.runToolCommand('yt-dlp', ['--version']),
      ]);

    return {
      ffmpegAvailable: !ffmpegWhich.startsWith('Command failed'),
      ytDlpAvailable: !ytDlpWhich.startsWith('Command failed'),
      ffmpegVersion: ffmpegVersion.split('\n')[0],
      ytDlpVersion: ytDlpVersion.split('\n')[0],
    };
  }

  @Post('generate-reviewed')
  @ApiOperation({
    operationId: 'aiGenerateReviewed',
    summary: 'Generate reviewed AI question drafts without saving them',
    description:
      'Generates multiple-choice question drafts from the configured AI provider, validates quality issues, and returns drafts only.',
  })
  @ApiBody({
    type: GenerateReviewedQuestionsDto,
    examples: {
      default: {
        summary: 'Generate reviewed Arabic World Cup questions',
        value: {
          catalogName: 'رياضة',
          categoryName: 'كأس العالم',
          difficulty: 'medium',
          count: 10,
          language: 'ar',
        },
      },
    },
  })
  @ApiResponse({
    status: 201,
    type: GenerateReviewedQuestionsResponseDto,
    description: 'Reviewed question drafts generated successfully',
  })
  async generateReviewed(@Body() body: GenerateReviewedQuestionsDto) {
    const result = await this.aiAgentService.generateReviewedQuestions(body);

    return {
      statusCode: HttpStatus.CREATED,
      ...result,
    };
  }

  @Post('save-drafts')
  @ApiOperation({
    operationId: 'aiSaveReviewedDrafts',
    summary: 'Persist reviewed questions as unpublished AI drafts',
  })
  @ApiResponse({ status: 201, type: SaveReviewedDraftsResponseDto })
  async saveDrafts(@Body() body: SaveReviewedDraftsDto) {
    const result = await this.aiAgentService.saveReviewedDrafts(body);
    return {
      ...result,
      savedQuestions: QuestionResponseMapper.toResponseList(
        result.savedQuestions,
      ),
    };
  }

  private async runToolCommand(
    command: string,
    args: string[],
  ): Promise<string> {
    try {
      const { stdout, stderr } = await execFileAsync(command, args);
      return [stdout, stderr].filter(Boolean).join('\n').trim();
    } catch (error) {
      if (error instanceof Error) {
        return error.message;
      }

      return String(error);
    }
  }
}