# Lammah-game-frontend/src/api/generated/models/questionAssetResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lammah Game API
 * Runtime HTTP contract for Lammah game administration, content, and gameplay.
 * OpenAPI spec version: 1.0.0
 */
import type { QuestionAssetType } from "./questionAssetType";
import type { QuestionAssetResponseDtoMetadata } from "./questionAssetResponseDtoMetadata";

export interface QuestionAssetResponseDto {
  type: QuestionAssetType;
  url: string;
  source: string;
  sourceUrl?: string;
  searchQuery?: string;
  provider?: string;
  duration?: number;
  metadata?: QuestionAssetResponseDtoMetadata;
}