# Lammah-game-frontend/src/api/generated/models/questionResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lammah Game API
 * Runtime HTTP contract for Lammah game administration, content, and gameplay.
 * OpenAPI spec version: 1.0.0
 */
import type { DifficultyLevel } from "./difficultyLevel";
import type { QuestionResponseDtoPoints } from "./questionResponseDtoPoints";
import type { QuestionResponseDtoScore } from "./questionResponseDtoScore";
import type { GameMode } from "./gameMode";
import type { QuestionStatus } from "./questionStatus";
import type { QuestionSource } from "./questionSource";
import type { QuestionResponseDtoPrimaryAsset } from "./questionResponseDtoPrimaryAsset";
import type { QuestionResponseDtoCoverImage } from "./questionResponseDtoCoverImage";
import type { QuestionResponseDtoPrimaryAssetRequest } from "./questionResponseDtoPrimaryAssetRequest";
import type { QuestionResponseDtoCoverImageRequest } from "./questionResponseDtoCoverImageRequest";
import type { AssetStatus } from "./assetStatus";
import type { QuestionResponseDtoAssetFailureDiagnostics } from "./questionResponseDtoAssetFailureDiagnostics";
import type { QuestionResponseDtoGameplayMetadata } from "./questionResponseDtoGameplayMetadata";
import type { QuestionResponseDtoAiMetadata } from "./questionResponseDtoAiMetadata";
import type { QuestionResponseDtoMetadata } from "./questionResponseDtoMetadata";

export interface QuestionResponseDto {
  _id: string;
  id?: string;
  category?: string;
  categoryId?: string;
  question: string;
  answer?: string;
  correctAnswer?: string;
  wrongAnswers: string[];
  explanation?: string;
  difficulty: DifficultyLevel;
  points?: QuestionResponseDtoPoints;
  score?: QuestionResponseDtoScore;
  gameMode?: GameMode;
  type?: string;
  status: QuestionStatus;
  source: QuestionSource;
  /** @nullable */
  primaryAsset?: QuestionResponseDtoPrimaryAsset;
  /** @nullable */
  coverImage?: QuestionResponseDtoCoverImage;
  /** @nullable */
  primaryAssetRequest?: QuestionResponseDtoPrimaryAssetRequest;
  /** @nullable */
  coverImageRequest?: QuestionResponseDtoCoverImageRequest;
  coverImageStatus?: AssetStatus;
  coverImageFailureReason?: string;
  assetStatus?: AssetStatus;
  assetFailureReason?: string;
  assetFailureStep?: string;
  assetFailureDiagnostics?: QuestionResponseDtoAssetFailureDiagnostics;
  gameplayMetadata?: QuestionResponseDtoGameplayMetadata;
  aiMetadata?: QuestionResponseDtoAiMetadata;
  metadata?: QuestionResponseDtoMetadata;
  mediaUrl?: string;
  hasPreviewAudio?: boolean;
  isFreeGameQuestion?: boolean;
  qualityScore?: number;
  issues?: string[];
  createdAt?: string;
  updatedAt?: string;
}